[codex] add offline rut-simulate binary (#22)
* add offline rut-simulate binary

* fix simulate PR CI

* fix clang-tidy errors for simulate

* fix remaining clang-tidy arena warning

* fix clang-tidy alloc_array template

* suppress clang-tidy false positives in arena

* address simulate review feedback

* expand simulate regression coverage

* tighten simulate manifest validation

* fix simulate review issues

* harden simulate failure contracts

* polish simulate review feedback

* expand simulate cli regression tests

* handle truncated simulate captures

* expand truncated capture test matrix

* tighten simulate summary formatting

* wire simulate test binary dependency

* lock simulate engine init review regressions

* fix simulate engine review issues

+++ include/rut/jit/handler_abi.h
@@ -103,7 +103,7 @@ struct HandlerCtx {
     T load_slot(u32 idx) const {
         static_assert(sizeof(T) <= 8, "Slot values must be <= 8 bytes");
         T val{};
-        const u8* src = slots() + idx * 8;
+        const u8* src = slots() + static_cast<size_t>(idx) * 8;
         __builtin_memcpy(&val, src, sizeof(T));
         return val;
     }
@@ -111,7 +111,7 @@ struct HandlerCtx {
     template <typename T>
     void store_slot(u32 idx, T val) {
         static_assert(sizeof(T) <= 8, "Slot values must be <= 8 bytes");
-        u8* dst = slots() + idx * 8;
+        u8* dst = slots() + static_cast<size_t>(idx) * 8;
         u64 zero = 0;
         __builtin_memcpy(dst, &zero, 8);
         __builtin_memcpy(dst, &val, sizeof(T));
